requirements.txt
PyGithub>=2.1.1
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
httpx>=0.25.0
